Seperate loop and key checks.
For some reason I had an enviroment who wasn't creating new
Invoice line items.  After some digging I found that the issue
lied with adding the check inside of the for..in loop.  You can
recreate this issue inside of Chrome console with the following
code:

var obj = { a: 1, b: 2, c: 3 };
for(var key in obj && key != 'a') { console.log(key, obj[key]); }

Changing the check to be inside the loop fixes this issue.